# Comprehensive performance validation and optimization (Issues #72, #77)